combine bsm calc tests into 1 file
